Move OpenShooterCharacter to Character dir
