+++ Source/OpenShooter/Private/Character/OpenShooterCharacter.cpp
@@ -1,6 +1,6 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
-#include "OpenShooterCharacter.h"
+#include "Character/OpenShooterCharacter.h"
 
 #include "Camera/CameraComponent.h"
 #include "Components/CapsuleComponent.h"
